mpidig: separate generic code from am specific code
Some functionalities are AM specific (e.g., registering AM callbacks)
and thus should not be placed in the generic code. This commit splits
these into separate files and directories.
