+++ src/mpid/ch4/generic/Makefile.mk
@@ -6,3 +6,9 @@
 ##
 
 include $(top_srcdir)/src/mpid/ch4/generic/am/Makefile.mk
+
+AM_CPPFLAGS += -I$(top_srcdir)/src/mpid/ch4/generic
+
+noinst_HEADERS += src/mpid/ch4/generic/mpidig.h
+
+mpi_core_sources += src/mpid/ch4/generic/mpidig_init.c
