+++ src/mpid/ch4/generic/am/Makefile.mk
@@ -12,10 +12,10 @@
 
 AM_CPPFLAGS += -I$(top_srcdir)/src/mpid/ch4/generic/am
 
-noinst_HEADERS += src/mpid/ch4/generic/am/mpidig_send.h \
-                  src/mpid/ch4/generic/am/mpidig_recv.h \
-                  src/mpid/ch4/generic/am/mpidig_msg.h \
-                  src/mpid/ch4/generic/am/mpidig.h
+noinst_HEADERS += src/mpid/ch4/generic/am/mpidig_am_send.h \
+                  src/mpid/ch4/generic/am/mpidig_am_recv.h \
+                  src/mpid/ch4/generic/am/mpidig_am_msg.h \
+                  src/mpid/ch4/generic/am/mpidig_am.h
 
-mpi_core_sources += src/mpid/ch4/generic/am/mpidig_globals.c \
-                    src/mpid/ch4/generic/am/mpidig_init.c
+mpi_core_sources += src/mpid/ch4/generic/am/mpidig_am_globals.c \
+                    src/mpid/ch4/generic/am/mpidig_am_init.c
