+++ src/mpid/ch4/generic/am/mpidig_am.h
@@ -8,10 +8,8 @@
  *  to Argonne National Laboratory subject to Software Grant and Corporate
  *  Contributor License Agreement dated February 8, 2012.
  */
-#ifndef MPIDIG_H_INCLUDED
-#define MPIDIG_H_INCLUDED
-
-#include <mpidimpl.h>
+#ifndef MPIDIG_AM_H_INCLUDED
+#define MPIDIG_AM_H_INCLUDED
 
 #define MPIDI_AM_HANDLERS_MAX (64)
 
@@ -34,9 +32,8 @@ extern MPIDIG_global_t MPIDIG_global;
 void MPIDIG_am_reg_cb(int handler_id,
                       MPIDIG_am_origin_cb origin_cb, MPIDIG_am_target_msg_cb target_msg_cb);
 int MPIDIG_am_reg_cb_dynamic(MPIDIG_am_origin_cb origin_cb, MPIDIG_am_target_msg_cb target_msg_cb);
-int MPIDIG_init(void);
-void MPIDIG_finalize(void);
 
-int MPIDIG_comm_abort(MPIR_Comm * comm, int exit_code);
+int MPIDIG_am_init(void);
+void MPIDIG_am_finalize(void);
 
-#endif /* MPIDIG_H_INCLUDED */
+#endif /* MPIDIG_AM_H_INCLUDED */
