+++ src/mpid/ch4/generic/am/mpidig_am_msg.h
@@ -3,8 +3,8 @@
  *  (C) 2020 by Argonne National Laboratory.
  *      See COPYRIGHT in top-level directory.
  */
-#ifndef MPIDIG_MSG_H_INCLUDED
-#define MPIDIG_MSG_H_INCLUDED
+#ifndef MPIDIG_AM_MSG_H_INCLUDED
+#define MPIDIG_AM_MSG_H_INCLUDED
 
 /* This file is for supporting routines used for generic layer message matching
  * and data transfer. They are used by protocols such as send, ssend, and send_long.
@@ -147,4 +147,4 @@ MPL_STATIC_INLINE_PREFIX int MPIDIG_recv_copy_seg(void *payload, MPI_Aint payloa
     }
 }
 
-#endif /* MPIDIG_MSG_H_INCLUDED */
+#endif /* MPIDIG_AM_MSG_H_INCLUDED */
