+++ src/mpid/ch4/generic/mpidig.h
@@ -1,0 +1,22 @@
+/* -*- Mode: C; c-basic-offset:4 ; indent-tabs-mode:nil ; -*- */
+/*
+ *  (C) 2006 by Argonne National Laboratory.
+ *      See COPYRIGHT in top-level directory.
+ *
+ *  Portions of this code were written by Intel Corporation.
+ *  Copyright (C) 2011-2016 Intel Corporation.  Intel provides this material
+ *  to Argonne National Laboratory subject to Software Grant and Corporate
+ *  Contributor License Agreement dated February 8, 2012.
+ */
+#ifndef MPIDIG_H_INCLUDED
+#define MPIDIG_H_INCLUDED
+
+#include <mpidimpl.h>
+#include "mpidig_am.h"
+
+int MPIDIG_init(void);
+void MPIDIG_finalize(void);
+
+int MPIDIG_comm_abort(MPIR_Comm * comm, int exit_code);
+
+#endif /* MPIDIG_H_INCLUDED */
